requirements.txt
discord.py==2.3.2
python-dotenv
asyncpg
audioop-lts; python_version >= "3.13"

